move from Details into KeyValue class
