+++ core/src/main/java/com/graphhopper/search/EdgeKVStorage.java
@@ -429,15 +429,12 @@ public class EdgeKVStorage {
         return vals.getCapacity() + keys.getCapacity();
     }
 
-    // TODO move into KeyValue class?
-    public static class Details {
+    public static class KeyValue {
         public static final String STREET_NAME = "street_name";
         public static final String STREET_REF = "street_ref";
         public static final String STREET_DESTINATION = "street_destination";
         public static final String STREET_DESTINATION_REF = "street_destination_ref";
-    }
 
-    public static class KeyValue {
         public String key;
         public Object value;
         public boolean fwd, bwd;
