+++ core/src/main/java/com/graphhopper/storage/BaseGraph.java
@@ -974,7 +974,7 @@ public class BaseGraph implements Graph, Closeable {
 
         @Override
         public String getName() {
-            String name = (String) getValue(EdgeKVStorage.Details.STREET_NAME);
+            String name = (String) getValue(EdgeKVStorage.KeyValue.STREET_NAME);
             // preserve backward compatibility (returns empty string if name tag missing)
             return name == null ? "" : name;
         }
