+++ core/src/test/java/com/graphhopper/routing/PathTest.java
@@ -35,11 +35,10 @@ import org.junit.jupiter.api.Test;
 
 import java.util.*;
 
-import static com.graphhopper.search.EdgeKVStorage.Details.STREET_NAME;
+import static com.graphhopper.search.EdgeKVStorage.KeyValue.STREET_NAME;
 import static com.graphhopper.search.EdgeKVStorage.KeyValue.createKV;
 import static com.graphhopper.storage.AbstractGraphStorageTester.assertPList;
 import static com.graphhopper.util.Parameters.Details.*;
-import com.graphhopper.search.EdgeKVStorage.KeyValue;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
