+++ core/src/test/java/com/graphhopper/storage/AbstractGraphStorageTester.java
@@ -32,7 +32,7 @@ import org.junit.jupiter.api.Test;
 
 import java.io.File;
 
-import static com.graphhopper.search.EdgeKVStorage.Details.STREET_NAME;
+import static com.graphhopper.search.EdgeKVStorage.KeyValue.STREET_NAME;
 import static com.graphhopper.search.EdgeKVStorage.KeyValue.createKV;
 import static org.junit.jupiter.api.Assertions.*;
