+++ web/src/test/java/com/graphhopper/application/resources/ExtendedJsonResponseTest.java
@@ -89,7 +89,7 @@ public class ExtendedJsonResponseTest {
         pointList.add(-3.4445, -38.9990);
         pointList.add(-3.5550, -38.7990);
         return new VirtualEdgeIteratorState(0, 0, 0, 1, 10, new IntsRef(1),
-                EdgeKVStorage.KeyValue.createKV(EdgeKVStorage.Details.STREET_NAME, "test of iterator"), pointList, false);
+                EdgeKVStorage.KeyValue.createKV(EdgeKVStorage.KeyValue.STREET_NAME, "test of iterator"), pointList, false);
     }
 
 }
